Create the basic structure of a c++ project for solving the sokoban.
It is compiled with CMake, and the CMakeLists file is included. Also,
an example map is included in the resources folder.

A main script is included. It reads a text file with the map, extracts
the map parameters (width, height and number of diamonds), and
analyzes which map squares are walkable, creating a boolean array.