Add class methods for updating the puzzle state
This is achieved with 2 different methods: One for updating the
position of the player, and the other for updating the position of a
box or diamond.

The instructions in the main script for reading the data in the text
file and creating the puzzle instance have been moved to a separate
routine in the same file.