Add class method for performing a movement in the puzzle
The method uses the previously calculated action, and checks that the
movement is possible. If so, updates the position of the man/robot and
the box/diamond, in case that there is one.

It has been prepared the usage of a new method, 'is_repeated_state',
that must check if the solver passed through the new state in a
previous stage.