Add a deadlocks map, and update it before de main loop starts
The map is generated through a method that is call after the other
maps are created. It looks for the corners in the map, and for the
squares connecting them, withouth a scape route between them. Goal
squares are excluded from the analysis.

Besides, a bug in the deadlock_state function was fixed: A break
statement was preventing from checking if the boxes are on a deadlock,
in case that other one is over a goal. This had to be changed by a
continue statement