Change the actions variable to a vector of vectors and keep historic
This way, the valid actions for each state are stored even after
passing through this state, for using it if in the future the solver
goes back to the same state.

The historical values of the chosen actions are stored, and are
printed at the end of the execution, in case that the solver finds a
suitable solution.