Comment methods to be replaced and add labeling algorithm
The new state representation will consist in the position of all the
boxes, plus the region of the player/man. At each iteration, instead
of evaluating all the possible moves of the player, they are
evaluated all the boxes that can be moved.

Firstly, a 4-connectivity algorithm has been developed, which will be
used later for knowing in which region the player is, and which boxes
can be pushed, and where.